backend/requirements.txt
django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.4
python-dotenv==1.0.0